refactor: #11 성능 최적화를 위한 애니메이션 간소화 및 클라이언트 컴포넌트 분리리
